+++ CHANGELOG.md
@@ -4,6 +4,10 @@ HyBIG follows semantic versioning. All notable changes to this project will be
 documented in this file. The format is based on [Keep a
 Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+* [GITC-8962](https://bugs.earthdata.nasa.gov/browse/GITC-8962): Memory and performance optimization for reading data from source when width and height or scale size parameters are not the full resolution of the data.
+
 ## [v2.6.2] - 2026-04-14
 
 ### Changed